// src/features/models/moduleHotelConfig.ts
export interface ModuleHotelConfiguration {
  name: string
  spokenLanguages: string[]
  stars: number
  images: string[]
  locationCoordinates: string[]
  checkinHour: string
  checkoutHour: string
  startDate: Date
  endDate: Date
  rooms: string
  breakfast: string
  wifi: string
  address: string
  phoneNumber: string
  contactEmail: string
}